Squashed commit of the following:
commit 27474bf1cc5a0963f0882ab56bc3652631b6454f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 22:09:00 2026 -0600

    fixes

commit 10b6e39487eeab740eb06d9fc2c96b98d99309c2
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 21:22:14 2026 -0600

    tests+shims

commit a504958ff00ab41309352b729f94cc41341b4d7c
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 21:04:49 2026 -0600

    Update README.md

commit 1a682d783bbb7bb136d7e150d8d56c5b9807d5e1
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 20:43:20 2026 -0600

    compile fixes

commit d532805c6323678d5b4392224b9a8159d110882b
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 19:03:33 2026 -0600

    tests

commit d5bd54db483112a07dc69ff6549bfe247182237a
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 18:33:57 2026 -0600

    tests

commit 6d644b518aa412d798ff8035bd7cec56b3d73da5
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 18:02:09 2026 -0600

    fixes

commit b2cd2a738dd2c40b0e7630484a4676dd05fd4638
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 17:10:09 2026 -0600

    fixes

commit bfd23d11995df5b4738b1f0ec82cf0e95bc828b5
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 16:37:15 2026 -0600

    fixes

commit 9f2c09251580d5a3218ae12fab7f7359a5e75b8c
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 13:19:27 2026 -0600

    fixes

commit 0a2531374109450b5b10d1234a3767ae54ce56d1
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 12:12:55 2026 -0600

    fractional timezone

commit 6ec6a7f02987bb4d0796869b76f4ad7f79871121
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 10:03:31 2026 -0600

    Update HydroInputDrivers.h

commit 9115e61e61a5dade5287e9a75fc231d6d3770d03
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 20:41:46 2026 -0600

    fixes

commit 88d38f0f0509c1c10ca70b5edbe64081604cd9f9
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 18:25:39 2026 -0600

    fixes

commit 97904a236cfa12183bf97f4483dfc1c5e9efa77e
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 17:46:31 2026 -0600

    Update Hydruino.cpp

commit 0e3620f4528e83062c5f7ea8c340d5d34243b060
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 15:28:00 2026 -0600

    Update README.md

commit 657b9bfa63e8109eaea39136eaeab907b03abe63
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 14:59:27 2026 -0600

    updates

commit f1d9b02afd257b46ff9d4bc51e0cfcd3273214ca
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 23:18:30 2026 -0600

    Update HydroStrings.h

commit 2f49c9852fb84c5a8272aa69d055b0a910784ad4
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 21:44:07 2026 -0600

    Update HydroStrings.cpp

commit 3d63278d883e3940c87040b84ed6a42801bbecf0
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 13:53:31 2026 -0600

    fixes

commit 338fa0ed279a4a5580ac2f7742aa4fe78e110174
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 28 22:00:46 2026 -0600

    Update HydroCoreLogic.h

commit 40e82ecaa8a545bf42908969dc8fc5b2c1022003
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 28 15:35:27 2026 -0600

    removing validate crap

commit 74cddaad6996734a8c956dc59dd88fbcfab67d92
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Aug 25 13:32:09 2026 -0600

    Update HydroAttachments.hpp

commit f4f94afc91f56d57e909e9f248a8d454a26646ee
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 17:43:30 2026 -0600

    Update HydroCropsLibrary.h

commit ee2480fd8d9a9d74107cccb6be717404c636e85b
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 16:22:20 2026 -0600

    Update HydroInterfaces.hpp

commit 06b61ded42ca0332f5a52dbc7276a59662404cbc
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 15:28:30 2026 -0600

    Update HydroCallback.hh

commit a2593f66a78089631c58ba61434d58890bd3993c
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 13:50:34 2026 -0600

    fixing main hpp

commit 566e6c0d7fe5f65b7e16521231759cb41d3b908f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 21 00:29:17 2026 -0600

    Update HydroModules.cpp

commit 6c7fc696f382fe30d25bbb866963f8b8d30a62b4
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Thu Aug 20 22:14:01 2026 -0600

    Update README.md

commit 015799d10a651dcdf79039b06a77baf186fc6333
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Thu Aug 20 21:37:52 2026 -0600

    upping version

commit cdf9142d83e0e87f5da44546226fd98eaf270736
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Thu Aug 20 19:31:27 2026 -0600

    Removing old stuff & consistency

+++ src/HydroStreams.h
@@ -36,7 +36,7 @@ public:
 
 protected:
     I2C_eeprom *_eeprom;
-    uint16_t _readAddress, _writeAddress, _endAddress;
+    uint32_t _readAddress, _writeAddress, _endAddress;
 };
 
 
@@ -63,7 +63,7 @@ protected:
 
 class HydroWiFiStorageFileStream : public Stream {
 public:
-    HydroWiFiStorageFileStream(WiFiStorageFile file, uintptr_t seekPos = 0);
+    HydroWiFiStorageFileStream(WiFiStorageFile file, uint32_t seekPos = 0);
     virtual ~HydroWiFiStorageFileStream();
 
     virtual int available() override;
@@ -81,9 +81,9 @@ protected:
     WiFiStorageFile _file;
     uint8_t _buffer[HYDRO_WIFISTREAM_BUFFER_SIZE];
     size_t _bufferOffset;
-    uintptr_t _bufferFileOffset;
+    uint32_t _bufferFileOffset;
     WiFiStorageFileDirection _bufferDirection;
-    uintptr_t _readOffset, _writeOffset, _endOffset;
+    uint32_t _readOffset, _writeOffset, _endOffset;
 
     void prepareReadBuffer();
     void prepareWriteBuffer();
